tests/requirements.txt
ledger_app_clients.ethereum
pytest
ragger[speculos,ledgercomm]